all tests

+++ src/main/java/ua/nure/lukianova/SummaryTask4/web/command/DeleteTestCommand.java
@@ -1,17 +1,17 @@
 package ua.nure.lukianova.SummaryTask4.web.command;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ua.nure.lukianova.SummaryTask4.exception.AppException;
 import ua.nure.lukianova.SummaryTask4.service.TestService;
 import ua.nure.lukianova.SummaryTask4.web.Path;
+import ua.nure.lukianova.SummaryTask4.web.Parameter;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-
-import static ua.nure.lukianova.SummaryTask4.web.Parameter.TEST_ID;
 
 public class DeleteTestCommand extends Command {
 
@@ -26,10 +26,13 @@ public class DeleteTestCommand extends Command {
 
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException, AppException {
-        long testId = Long.valueOf(request.getParameter(TEST_ID));
+
+        if (StringUtils.isEmpty(request.getParameter(Parameter.TEST_ID))) {
+            throw new AppException("Invalid input");
+        }
+        long testId = Long.valueOf(request.getParameter(Parameter.TEST_ID));
         testService.delete(testId);
         return Path.COMMAND_LIST_TESTS;
     }
-
 
 }
