+++ src/main/java/ua/nure/lukianova/SummaryTask4/web/command/DeleteUserCommand.java
@@ -1,17 +1,17 @@
 package ua.nure.lukianova.SummaryTask4.web.command;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ua.nure.lukianova.SummaryTask4.exception.AppException;
 import ua.nure.lukianova.SummaryTask4.service.UserService;
+import ua.nure.lukianova.SummaryTask4.web.Parameter;
 import ua.nure.lukianova.SummaryTask4.web.Path;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-
-import static ua.nure.lukianova.SummaryTask4.web.Parameter.USER_ID;
 
 public class DeleteUserCommand extends Command {
 
@@ -25,7 +25,11 @@ public class DeleteUserCommand extends Command {
 
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException, AppException {
-        long userId = Long.valueOf(request.getParameter(USER_ID));
+
+        if(StringUtils.isEmpty(request.getParameter(Parameter.USER_ID))){
+            throw new AppException("Invalid input");
+        }
+        long userId = Long.valueOf(request.getParameter(Parameter.USER_ID));
         userService.delete(userId);
         return Path.COMMAND_LIST_USERS;
     }
